rpo_mojoco/rpo/requirements.txt
matplotlib=3.1.3
numpy=1.18.1
pip=20.0.2
python=3.7.7
scipy=1.4.1
seaborn=0.10.1
tensorflow-estimator=2.1.0
tensorflow-gpu=2.1.0
gym==0.17.1
mujoco-py==1.50.1.68
